Smart-Surveillance-System-for-Vehicle-Crash-Detection/requirements_fixed.txt
# Fixed requirements for Python 3.14 compatibility
# Install NumPy 1.x first, then other packages

numpy==1.26.4
opencv-python-headless
ultralytics
albumentations
pyyaml
flask
flask-cors
requests
tqdm
pandas
torch
torchvision
torchaudio

